requirements.txt
# Core dependencies
streamlit==1.36.0
crewai==0.28.5
langchain-groq==0.1.1
groq==0.4.1
langchain-core==0.1.53
langchain-community==0.0.38
langchain-openai==0.1.3
python-dotenv==1.0.1
reportlab==4.1.0

# PDF processing
PyPDF2==3.0.1
PyMuPDF==1.24.6

# Data processing
pandas==2.2.0
numpy==1.26.4
semantic-version==2.10.0

# Visualization
matplotlib==3.8.3
seaborn==0.13.2 